Move extract_function_names back to programs_database
